requirements.txt
python-dotenv==0.19.0
discord.py==1.7.3
pymongo==3.12.0
beautifulsoup4==4.10.0
requests==2.26.0
Pillow==8.3.2
selenium==3.141.0